refactor: move convert module into its own file
